I init MediaUpload folder.
