+++ components/MediaUpload/index.tsx
@@ -1,0 +1,3 @@
+import MediaUpload from './MediaUpload'
+
+export default MediaUpload
